Admit scatter: point model, numeric x-scale, and full cross-language parity
Implements REQ-CHART-002 (DEC-015): scatter as the next certified chart
type, landing two genuine generalizations shared across every chart
type rather than an isolated marks function like bar was.

Point model (series[].data element normalizes to a bare number
x=index, positional [x,y], or object {x,y}): registered identically in
the schema (type-conditional oneOf/allOf, scatter only), the Python
validator, and the Go validator, with a custom Series.UnmarshalJSON in
Go and a chart-type-gated parse branch in Python's from_dict. Added
Series.data_points ([]Datum in Go) as a new field, unread by every
existing renderer, so line/column/area/bar's Series.data path is
completely untouched.

Numeric linear x-scale (_cartesian.py/.go): a third x-scale mode
alongside point/band, reusing nice_ticks/fmt_num verbatim with
include_zero=False and the same degenerate-domain guard value_pix
already had for y. Landed as strictly additive branches (new `if
scale == "linear"` checks) in build_frame/buildFrame and the chrome
head's x-axis rendering - every existing branch for point/band scale
is byte-for-byte the same code that ran before scatter existed.

Byte-identity gate (REQ-CHART-002's acceptance #3): verified
immediately after the frame/spec changes landed, before writing any
scatter-specific code - full Python and Go test suites green with zero
changes to any line/column/area/bar golden. Then built scatter.py/
scatter.go (reusing line's marker builder exactly as area.py already
does), 6 golden fixtures (basic/correlation/regression/themed-dark/
adversarial/xy-points - the last proving the positional/object forms
explicitly), invalid-fixtures.json for the new validation surface, and
verified full byte-parity: Python and Go produce byte-identical SVG
for all 6 fixtures, confirmed both by direct diff and by
check_direct_cross_render.py (29 examples now, zero diff).

Found and fixed two real bugs during the browser-qualification pass:
the typed Python construction path (ChartSpec(series=[Series(...)]),
bypassing from_dict) never normalized data_points, rendering zero
points; fixed with a ChartSpec.__post_init__. The symmetric Go gap (a
raw *ChartSpec struct literal bypassing FromJSON) is fixed with a
bare-number backfill in renderScatterSVG.

Generalized the accessible data table to a long-format (series, x, y)
row shape in both languages (render.py's _data_table, Go's new
scatterDataTable) rather than coercing points into the category-shaped
table. Added scatter to the benchmark workload matrix (positional
[x,y] data, exercising the new scale path, not just the fast path) and
to the runtime-browser CI suite (scatter-browser-qualification.test.js,
proving the numeric data-x contract and unconnected-points behavior
live in Chromium).

Closed GATE-S9 (0.0.0.3 Qualification): fresh re-run of the full
qualification suite, a recorded accessibility/security review
(SC-REL-018) with a live ARIA snapshot, and a recorded performance
baseline review (SC-REL-019) with clean (uncontended) benchmark runs
in both languages. REQ-CHART-002 moves approved -> verified.

Also fixed three unrelated stale status claims found along the way:
README's bar row still said "targeting 0.0.0.2" despite 0.0.0.2 having
shipped, and CHARTS.md's column/area/bar rows didn't reflect their
actual certified/shipped status.

+++ libs/go/column.go
@@ -56,7 +56,7 @@ func columnMarks(f *cartesianFrame, p *strings.Builder) {
 			if i >= f.n {
 				break
 			}
-			cxBand := f.xpix(i)
+			cxBand := f.xpix(float64(i))
 			var left, y, h float64
 			if stacked {
 				left = cxBand - groupW/2
